+++ HISTORY.rst
@@ -5,6 +5,22 @@ History
 
 As of 1.0.0, `semantic versioning <http://semver.org/>`__ is used.
 
+11.0.0
+++++++
+released 2017-12-09
+
+- breaking: list calls now default to max_results=None, increasing the default number of results from 100 to 999
+- add updated_after param to song/playlist listing to support differential updates
+- add support for free radio stations
+- add filepath+extension to unsupported file exception message
+- fix "I'm Feeling Lucky" station never refreshing its seed
+- fix crashes caused by some 503s during uploading
+- fix gmtools for https://github.com/simon-weber/Google-Music-Playlist-Importer
+- fix AAC and ALAC content type upload detection
+- blacklist requests 2.8.2
+- improve id documentation
+- update schemas
+
 10.1.2
 ++++++
 released 2017-04-03